refactored all IDs to be size_t rather than int

+++ include/APRIL/storage.h
@@ -7,7 +7,7 @@ namespace APRIL
 {
     namespace writer
     {
-        DB_STATUS saveAPRIL(FILE* pFileALL, FILE* pFileFULL, uint recID, uint sectionID, AprilDataT* aprilData);
+        DB_STATUS saveAPRIL(FILE* pFileALL, FILE* pFileFULL, size_t recID, uint sectionID, AprilDataT* aprilData);
     }
 
     namespace reader
